little refactoring

+++ src/main/java/com/santander/gofTest/Controller/Address/AddressController.java
@@ -3,6 +3,8 @@ package com.santander.gofTest.Controller.Address;
 import com.santander.gofTest.Models.Address.Address;
 import com.santander.gofTest.Repository.AddressRepo;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -15,25 +17,30 @@ public class AddressController {
     private AddressRepo addressRepo;
 
     @GetMapping(value = "/")
-    public List<Address> addresses() {
-        return addressRepo.findAll();
+    public ResponseEntity<List<Address>> addresses() {
+        return ResponseEntity.status(HttpStatus.OK).body(addressRepo.findAll());
     }
 
     @PostMapping(value = "/add")
-    public void add(@RequestBody Address address) {
+    public ResponseEntity<String> add(@RequestBody Address address) {
         addressRepo.save(address);
+        return ResponseEntity.status(HttpStatus.CREATED).build();
     }
 
     @PutMapping(value = "/update/{id}")
-    public void upd(@PathVariable long id, @RequestBody Address address) {
+    public ResponseEntity<String> upd(@PathVariable long id, @RequestBody Address address) throws RuntimeException {
+        if (addressRepo.findById(id).isEmpty()) throw new RuntimeException("address not found");
         Address a = addressRepo.findById(id).get();
         a.update(address);
         addressRepo.save(a);
+        return ResponseEntity.status(HttpStatus.CREATED).build();
     }
 
     @DeleteMapping(value = "/delete/{id}")
-    public void del(@PathVariable long id) {
+    public ResponseEntity<String> del(@PathVariable long id) throws RuntimeException {
+        if (addressRepo.findById(id).isEmpty()) throw new RuntimeException("address not found");
         addressRepo.deleteById(id);
+        return ResponseEntity.status(HttpStatus.ACCEPTED).build();
     }
 }
